ComparisonShopper bugs fixed
Fixed a bug on launch where the roomLabelLeft @IBOutlet was not connected to the UILabel. Another bug prevented crashed the app after unsafely unwrapping the house1 properties when they were nil. Finally, the method setUpRightSideUI method was not showing anything after the Add Right button was tapped because the alpha was set to 0.